renaming audio encoder for readability
