+++ src/diffusers/models/autoencoders/__init__.py
@@ -1,5 +1,5 @@
 from .autoencoder_asym_kl import AsymmetricAutoencoderKL
-from .autoencoder_cosmos3_avae import Cosmos3AVAEAudioTokenizer
+from .autoencoder_cosmos3_audio import Cosmos3AVAEAudioTokenizer
 from .autoencoder_dc import AutoencoderDC
 from .autoencoder_kl import AutoencoderKL
 from .autoencoder_kl_allegro import AutoencoderKLAllegro
